Add files via upload
The Bank Loan Analysis project aims to examine loan transactions and customer data to enhance financial services. By leveraging a comprehensive dataset of loan applications and approvals, we seek to uncover valuable insights.
-PURPOSE OF ANALYSIS 
To evaluate loan offerings and terms.
To assess the risk and benefits involved in securing a bank loan.
To provide insights for informed borrowing decisions.